Refactor org dashboard into components
