+++ src/components/OrgDashboard/index.js
@@ -1,0 +1,5 @@
+export * from './EventPanel';
+export * from './OrgInfo';
+export * from './OrgButtons';
+export * from './OrgPhoto';
+
