+++ src/components/index.js
@@ -1,12 +1,10 @@
 export * from './Event';
 export * from './EventList';
 export * from './Navigation';
-export * from './OrganizationInfo';
 export * from '../styled/StyledUploadImage';
 export * from './RecurringEvent';
 export * from './FooterDiv';
 export * from './HeaderDiv';
-export * from './OrganizationInfo';
 export * from './UploadImage';
 export * from './FilteredComponent';
 export * from './FilterTopbar';
